Exceptions

+++ src/main/java/com/academia/fortal/api_academia/service/AlunoService.java
@@ -3,8 +3,10 @@ package com.academia.fortal.api_academia.service;
 import com.academia.fortal.api_academia.domain.dto.AlunoDTO;
 import com.academia.fortal.api_academia.domain.entities.Aluno;
 import com.academia.fortal.api_academia.dozer.DozerConverter;
+import com.academia.fortal.api_academia.exceptions.CommonsException;
 import com.academia.fortal.api_academia.repository.AlunoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,10 +30,20 @@ public class AlunoService {
 
     public AlunoDTO findById(Long id){
         var entity = alunoRepository.findById(id);
+        if(entity.isEmpty()){
+            throw new CommonsException(HttpStatus.NOT_FOUND,
+                    "academia.fortal.aluno.notfound",
+                    "O aluno com o ID informado, não foi encontrado");
+        }
         return DozerConverter.parseObject(entity.get(), AlunoDTO.class);
     }
 
     public void delete(Long id){
+        if (!alunoRepository.existsById(id)) {
+            throw new CommonsException(HttpStatus.NOT_FOUND,
+                    "academia.fortal.aluno.notfound",
+                    "O aluno com o ID informado não foi encontrado");
+        }
         alunoRepository.deleteById(id);
     }
