+++ src/main/java/com/academia/fortal/api_academia/service/EquipamentoService.java
@@ -3,9 +3,11 @@ package com.academia.fortal.api_academia.service;
 import com.academia.fortal.api_academia.domain.dto.EquipamentoDTO;
 import com.academia.fortal.api_academia.domain.entities.Equipamento;
 import com.academia.fortal.api_academia.dozer.DozerConverter;
+import com.academia.fortal.api_academia.exceptions.CommonsException;
 import com.academia.fortal.api_academia.repository.EquipamentoRepository;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,10 +31,20 @@ public class EquipamentoService {
 
     public EquipamentoDTO findById(Long id){
         var entity = equipamentoRepository.findById(id);
+        if(entity.isEmpty()){
+            throw new CommonsException(HttpStatus.NOT_FOUND,
+                    "academia.fortal.equipamento.notfound",
+                    "O equipamento com o ID informado, não foi encontrado");
+        }
         return DozerConverter.parseObject(entity.get(), EquipamentoDTO.class);
     }
 
     public void delete(Long id){
+        if (!equipamentoRepository.existsById(id)) {
+            throw new CommonsException(HttpStatus.NOT_FOUND,
+                    "academia.fortal.equipamento.notfound",
+                    "O equipamento com o ID informado não foi encontrado");
+        }
         equipamentoRepository.deleteById(id);
     }
